+++ apps/backend/src/modules/contributors/contributor.service.ts
@@ -1,6 +1,13 @@
+import {
+	MIN_GIT_EMAILS_LENGTH_FOR_SPLIT,
+	PAGE_INDEX_OFFSET,
+} from "~/libs/constants/constants.js";
 import { ExceptionMessage } from "~/libs/enums/enums.js";
 import { HTTPCode } from "~/libs/modules/http/http.js";
-import { type Service } from "~/libs/types/types.js";
+import {
+	type PaginationQueryParameters,
+	type Service,
+} from "~/libs/types/types.js";
 
 import { ContributorEntity } from "./contributor.entity.js";
 import { type ContributorRepository } from "./contributor.repository.js";
@@ -12,6 +19,7 @@ import {
 	type ContributorMergeRequestDto,
 	type ContributorPatchRequestDto,
 	type ContributorPatchResponseDto,
+	type ContributorSplitRequestDto,
 } from "./libs/types/types.js";
 
 class ContributorService implements Service {
@@ -52,8 +60,13 @@ class ContributorService implements Service {
 		return item.toObject();
 	}
 
-	public async findAll(): Promise<ContributorGetAllResponseDto> {
-		const contributors = await this.contributorRepository.findAll();
+	public async findAll(
+		parameters: PaginationQueryParameters,
+	): Promise<ContributorGetAllResponseDto> {
+		const contributors = await this.contributorRepository.findAll({
+			page: parameters.page - PAGE_INDEX_OFFSET,
+			pageSize: parameters.pageSize,
+		});
 
 		return {
 			items: contributors.items.map((item) => {
@@ -67,6 +80,7 @@ class ContributorService implements Service {
 					})),
 				};
 			}),
+			totalItems: contributors.totalItems,
 		};
 	}
 
@@ -88,6 +102,27 @@ class ContributorService implements Service {
 					})),
 				};
 			}),
+			totalItems: contributors.items.length,
+		};
+	}
+
+	public async findAllWithoutPagination(): Promise<ContributorGetAllResponseDto> {
+		const contributors =
+			await this.contributorRepository.findAllWithoutPagination();
+
+		return {
+			items: contributors.items.map((item) => {
+				const contributor = item.toObject();
+
+				return {
+					...contributor,
+					gitEmails: contributor.gitEmails.map((gitEmail) => ({
+						email: gitEmail.email,
+						id: gitEmail.id,
+					})),
+				};
+			}),
+			totalItems: contributors.items.length,
 		};
 	}
 
@@ -162,6 +197,59 @@ class ContributorService implements Service {
 		}
 
 		return item.toObject();
+	}
+
+	public async split(
+		contributorId: number,
+		payload: ContributorSplitRequestDto,
+	): Promise<ContributorGetAllItemResponseDto | null> {
+		const currentContributorEntity =
+			await this.contributorRepository.find(contributorId);
+		const hasContributor = currentContributorEntity !== null;
+
+		if (!hasContributor) {
+			throw new ContributorError({
+				message: ExceptionMessage.CONTRIBUTOR_NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		const currentContributor = currentContributorEntity.toObject();
+
+		if (
+			currentContributor.gitEmails.length <= MIN_GIT_EMAILS_LENGTH_FOR_SPLIT
+		) {
+			throw new ContributorError({
+				message: ExceptionMessage.CONTRIBUTOR_SPLIT_SINGLE_EMAIL,
+				status: HTTPCode.CONFLICT,
+			});
+		}
+
+		const splitEmail = currentContributor.gitEmails.find(
+			({ id }) => id === payload.gitEmailId,
+		);
+		const hasSplittedEmail = splitEmail !== undefined;
+
+		if (!hasSplittedEmail) {
+			throw new ContributorError({
+				message: ExceptionMessage.CONTRIBUTOR_SPLIT_FAILED,
+				status: HTTPCode.CONFLICT,
+			});
+		}
+
+		try {
+			const splitContributor = await this.contributorRepository.split(
+				payload.gitEmailId,
+				payload.newContributorName,
+			);
+
+			return splitContributor.toObject();
+		} catch {
+			throw new ContributorError({
+				message: ExceptionMessage.CONTRIBUTOR_SPLIT_FAILED,
+				status: HTTPCode.CONFLICT,
+			});
+		}
 	}
 
 	public update(): ReturnType<Service["update"]> {