+++ apps/frontend/src/modules/notifications/notifications-api.ts
@@ -2,6 +2,7 @@ import { APIPath, ContentType } from "~/libs/enums/enums.js";
 import { BaseHTTPApi } from "~/libs/modules/api/api.js";
 import { type HTTP } from "~/libs/modules/http/http.js";
 import { type Storage } from "~/libs/modules/storage/storage.js";
+import { type PaginationQueryParameters } from "~/libs/types/types.js";
 
 import { NotificationsApiPath } from "./libs/enums/enums.js";
 import { type NotificationGetAllResponseDto } from "./libs/types/types.js";
@@ -17,9 +18,30 @@ class NotificationApi extends BaseHTTPApi {
 		super({ baseUrl, http, path: APIPath.NOTIFICATIONS, storage });
 	}
 
-	public async getAll(): Promise<NotificationGetAllResponseDto> {
+	public async getAll(
+		query: PaginationQueryParameters,
+	): Promise<NotificationGetAllResponseDto> {
 		const response = await this.load(
 			this.getFullEndpoint(NotificationsApiPath.ROOT, {}),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "GET",
+				query: {
+					page: String(query.page),
+					pageSize: String(query.pageSize),
+				},
+			},
+		);
+
+		return await response.json<NotificationGetAllResponseDto>();
+	}
+
+	public async getAllUnread(): Promise<
+		Pick<NotificationGetAllResponseDto, "items">
+	> {
+		const response = await this.load(
+			this.getFullEndpoint(NotificationsApiPath.UNREAD, {}),
 			{
 				contentType: ContentType.JSON,
 				hasAuth: true,
@@ -27,7 +49,7 @@ class NotificationApi extends BaseHTTPApi {
 			},
 		);
 
-		return await response.json<NotificationGetAllResponseDto>();
+		return await response.json<Pick<NotificationGetAllResponseDto, "items">>();
 	}
 
 	public async markAsRead(id: number): Promise<boolean> {