+++ apps/frontend/src/pages/contributors/libs/helpers/get-contributor-columns/get-contributor-columns.helper.tsx
@@ -1,11 +1,15 @@
+import { Tooltip } from "~/libs/components/components.js";
+import { MIN_GIT_EMAILS_LENGTH_FOR_SPLIT } from "~/libs/constants/constants.js";
 import { type TableColumn } from "~/libs/types/types.js";
 
 import { ContributorMenu } from "../../components/components.js";
 import { type ContributorRow } from "../../types/types.js";
+import styles from "./styles.module.css";
 
 const getContributorColumns = (actions: {
 	onEdit: (contributorId: number) => void;
 	onMerge: (contributorId: number) => void;
+	onSplit: (contributorId: number) => void;
 }): TableColumn<ContributorRow>[] => [
 	{
 		accessorKey: "name",
@@ -13,8 +17,17 @@ const getContributorColumns = (actions: {
 		size: 200,
 	},
 	{
-		accessorFn: (contributor: ContributorRow): string =>
-			contributor.gitEmails.join(", "),
+		cell: ({ row: { original: contributor } }): JSX.Element => {
+			const gitEmails = contributor.gitEmails.join(", ");
+
+			return (
+				<div className={styles["git-email-cell-wrapper"]}>
+					<Tooltip content={gitEmails} id={gitEmails}>
+						{gitEmails}
+					</Tooltip>
+				</div>
+			);
+		},
 		header: "Git Emails",
 		size: 300,
 	},
@@ -28,8 +41,12 @@ const getContributorColumns = (actions: {
 		cell: ({ row: { original: contributor } }) => (
 			<ContributorMenu
 				contributorId={contributor.id}
+				isSplitEnabled={
+					contributor.gitEmails.length > MIN_GIT_EMAILS_LENGTH_FOR_SPLIT
+				}
 				onEdit={actions.onEdit}
 				onMerge={actions.onMerge}
+				onSplit={actions.onSplit}
 			/>
 		),
 		header: "",