+++ apps/frontend/src/pages/project-access-management/libs/components/project-group-create-form/project-group-create-form.tsx
@@ -49,7 +49,7 @@ const ProjectGroupCreateForm = ({
 				projectId,
 			}}
 			onSubmit={onSubmit}
-			submitLabel="Submit"
+			submitLabel="Create"
 			validationSchema={projectGroupCreateValidationSchema}
 		/>
 	);