+++ packages/shared/src/modules/contributors/libs/enums/contributors-api-path.enum.ts
@@ -2,6 +2,7 @@ const ContributorsApiPath = {
 	$ID: "/:id",
 	MERGE_$ID: "/merge/:id",
 	ROOT: "/",
+	SPLIT_$ID: "/split/:id",
 } as const;
 
 export { ContributorsApiPath };