+++ packages/shared/src/modules/notifications/libs/enums/notifications-api-path.enum.ts
@@ -1,6 +1,7 @@
 const NotificationsApiPath = {
 	$ID: "/:id",
 	ROOT: "/",
+	UNREAD: "/unread",
 } as const;
 
 export { NotificationsApiPath };