+++ packages/shared/src/modules/notifications/notifications.ts
@@ -5,5 +5,6 @@ export {
 	type NotificationBulkCreateResponseDto,
 	type NotificationCreateRequestDto,
 	type NotificationGetAllItemResponseDto,
+	type NotificationGetAllRequestDto,
 	type NotificationGetAllResponseDto,
 } from "./libs/types/types.js";