+++ apps/frontend/CHANGELOG.md
@@ -1,5 +1,46 @@
 # Changelog
 
+## [1.48.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.47.0...frontend-v1.48.0) (2024-09-24)
+
+
+### Features
+
+* add infinite scroll and pagination to notificationspopover gf-458 ([#490](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/490)) ([9daf38b](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/9daf38b8ab392357d17919a0a6507e884d5a52be))
+* add projects select to analytics page gf-348 ([#470](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/470)) ([47d8620](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/47d8620dff09d6f21febd1cb74fa08fb53fd6fca))
+
+## [1.47.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.46.0...frontend-v1.47.0) (2024-09-23)
+
+
+### Features
+
+* Add an ability to split contributors gf-462 ([#477](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/477)) ([adbf2af](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/adbf2afd81ac81518c50b158ab11cd89768e7c8a))
+
+## [1.46.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.45.0...frontend-v1.46.0) (2024-09-23)
+
+
+### Features
+
+* the earliest date availible to be choosen on the date range on the analytics page should be 1 year from today gf-482 ([#488](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/488)) ([f08b073](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/f08b0733e296f8856983f462689a94a4f80dd855))
+
+
+### Bug Fixes
+
+* adjusted search for modal to remove double reload gf-455 ([#493](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/493)) ([d80d5c5](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/d80d5c5b201b7e44f82cc432e264e24bda39069a))
+
+## [1.45.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.44.1...frontend-v1.45.0) (2024-09-23)
+
+
+### Features
+
+* Add an activity chart to contributor items on project details page gf-420 ([#468](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/468)) ([38cf6eb](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/38cf6ebc535a40e117b68f18c53cbecb58ae356f))
+* add pagination to the contributors page gf-484 ([#491](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/491)) ([75e0b48](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/75e0b483381fc8704feddbe3a6bf20b01331b98d))
+
+
+### Bug Fixes
+
+* date input component gf-472 ([#487](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/487)) ([b607eb3](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/b607eb3c2914717c60791f8185c29fb43376b891))
+* submit button labels in project create and update forms gf-474 ([#486](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/486)) ([00d9ade](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/00d9adec9421c90be57bc5f9a0b80e2c91c63eff))
+
 ## [1.44.1](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.44.0...frontend-v1.44.1) (2024-09-23)